refactor(ui): inline one-caller app, picker, and route helpers
