--- ui/src/App.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-import Dashboard from './views/Dashboard.vue'
-</script>
-
-<template>
-  <Dashboard />
-</template>
